+++ spec/helpers/codelist_helper_spec.rb
@@ -227,7 +227,7 @@ RSpec.describe CodelistHelper, type: :helper do
     describe "#ispf_partner_country_options" do
       context "when ODA" do
         it "returns the ODA partner country details in a hash" do
-          options = helper.ispf_partner_country_options(is_oda: true)
+          options = helper.ispf_partner_country_options(oda: true)
 
           expect(options.length).to eq(14)
           expect(options.first.code).to eq("BR")
@@ -239,13 +239,23 @@ RSpec.describe CodelistHelper, type: :helper do
 
       context "when non-ODA" do
         it "returns the non-ODA partner country details in a hash" do
-          options = helper.ispf_partner_country_options(is_oda: false)
+          options = helper.ispf_partner_country_options(oda: false)
 
           expect(options.length).to eq(12)
           expect(options.first.code).to eq("CA")
           expect(options.first.name).to eq("Canada")
           expect(options.last.code).to eq("NONE")
           expect(options.last.name).to eq("None (exemption agreed)")
+        end
+
+        context "when excluding 'none'" do
+          it "returns all except 'none'" do
+            options = helper.ispf_partner_country_options(oda: false, allow_none: false)
+
+            expect(options.length).to eq(11)
+            expect(options.last.code).to eq("US")
+            expect(options.last.name).to eq("USA")
+          end
         end
       end
     end