+++ spec/models/ispf_oda_partner_country_spec.rb
@@ -1,0 +1,9 @@
+require "spec_helper"
+
+RSpec.describe IspfOdaPartnerCountry do
+  describe "IspfOdaPartnerCountry" do
+    it "inherits from IspfPartnerCountry" do
+      expect(IspfOdaPartnerCountry).to be < IspfPartnerCountry
+    end
+  end
+end